Add getters "len" and "rad" for CPS enhancement experiments
Also whitespace got auto-corrected

Also the FIXME comment for the new properties is important:

We should have proper whitelisting facilities in CPS
to overcome 'accidential' access to properties that weren't meant
to be used in CPS. Also, this could add a re-mapping of names
to getter methods (or getters). This would make the namespace
less crowded and allow a better names for he CPS user!